requirements.txt
requests
progress
psutil
bs4
Pillow>=4.2.0
python-telegram-bot==12.6.1 
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
aria2p>=0.9.0,<0.15.0
python-dotenv>=0.10
tenacity>=6.0.0
python-magic
beautifulsoup4
pyrogram
TgCrypto
youtube_dl
feedparser
natsort
heroku3
aiohttp
speedtest-cli
messages
js2py
lxml
telegraph
pytz
megasdkrestclient>=0.1.1,<1.0.0
googletrans==3.1.0a0
google-cloud-storage
emoji
html-telegraph-poster>=0.2.31
nekos.py
pynewtonmath
wget
wikipedia
jikanpy
tabulate
psycopg2-binary
carbonnow
removebg==0.4
python-arq
OpenCC
cryptography
aiofiles
gTTS
ffmpeg-python
pykeyboard
geopy
pybase64
mal-api
anime_downloader>=5.0.7
fake_useragent>=0.1.11
quotefancy
maigret
countryinfo
emoji-country-flag
spotdl